requirements.txt
Django
django-widget_tweaks
requests
djangorestframework
